refactor remote data classes, enhanced example
