+++ ucapi/remote.py
@@ -5,12 +5,13 @@ Remote entity definitions.
 :license: MPL-2.0, see LICENSE for more details.
 """
 
-from dataclasses import KW_ONLY, dataclass
+import dataclasses
 from enum import Enum
 from typing import Any
 
 from ucapi.api_definitions import CommandHandler
 from ucapi.entity import Entity, EntityTypes
+from ucapi.ui import DeviceButtonMapping, EntityCommand, UiPage
 
 
 class States(str, Enum):
@@ -54,14 +55,6 @@ class Options(str, Enum):
     USER_INTERFACE = "user_interface"
 
 
-@dataclass
-class EntityCommand:
-    """Remote command definition for a button mapping or UI page definition."""
-
-    cmd_id: str
-    params: dict[str, str | int | list[str]] | None = None
-
-
 def create_send_cmd(
     command: str,
     delay: int | None = None,
@@ -69,7 +62,7 @@ def create_send_cmd(
     hold: int | None = None,
 ) -> EntityCommand:
     """
-    Create a send command.
+    Create a remote send command.
 
     :param command: command to send.
     :param delay: optional delay in milliseconds after the command or between repeats.
@@ -93,7 +86,7 @@ def create_sequence_cmd(
     repeat: int | None = None,
 ) -> EntityCommand:
     """
-    Create a sequence command.
+    Create a remote send sequence command.
 
     :param sequence: list of simple commands.
     :param delay: optional delay in milliseconds between the commands in the sequence.
@@ -108,175 +101,16 @@ def create_sequence_cmd(
     return EntityCommand(Commands.SEND_CMD_SEQUENCE.value, params)
 
 
-class Buttons(str, Enum):
-    """Physical buttons."""
-
-    BACK = "BACK"
-    HOME = "HOME"
-    VOICE = "VOICE"
-    VOLUME_UP = "VOLUME_UP"
-    VOLUME_DOWN = "VOLUME_DOWN"
-    MUTE = "MUTE"
-    DPAD_UP = "DPAD_UP"
-    DPAD_DOWN = "DPAD_DOWN"
-    DPAD_LEFT = "DPAD_LEFT"
-    DPAD_RIGHT = "DPAD_RIGHT"
-    DPAD_MIDDLE = "DPAD_MIDDLE"
-    GREEN = "GREEN"
-    YELLOW = "YELLOW"
-    RED = "RED"
-    BLUE = "BLUE"
-    CHANNEL_UP = "CHANNEL_UP"
-    CHANNEL_DOWN = "CHANNEL_DOWN"
-    PREV = "PREV"
-    PLAY = "PLAY"
-    NEXT = "NEXT"
-    POWER = "POWER"
-
-
-@dataclass
-class DeviceButtonMapping:
-    """Physical button mapping."""
-
-    button: str
-    short_press: EntityCommand | None = None
-    long_press: EntityCommand | None = None
-
-
-def create_btn_mapping(
-    button: Buttons,
-    short: str | EntityCommand | None = None,
-    long: str | EntityCommand | None = None,
-) -> DeviceButtonMapping:
-    """
-    Create a physical button command mapping.
-
-    :param button: physical button identifier.
-    :param short: associated short-press command to the physical button.
-                  A string parameter corresponds to a simple command, whereas an
-                  ``EntityCommand`` allows to customize the command.
-    :param long: associated long-press command to the physical button
-    :return: the created DeviceButtonMapping
-    """
-    if isinstance(short, str):
-        short = EntityCommand(short)
-    if isinstance(long, str):
-        long = EntityCommand(long)
-    return DeviceButtonMapping(button.value, short_press=short, long_press=long)
-
-
-@dataclass
-class Size:
-    """Item size in the button grid. Default size if not specified: 1x1."""
-
-    width: int = 1
-    height: int = 1
-
-
-@dataclass
-class Location:
-    """Button placement in the grid with 0-based coordinates."""
-
-    x: int
-    y: int
-
-
-@dataclass
-class UiItem:
-    """
-    A user interface item is either an icon or text.
-
-    - Icon and text items can be static or linked to a command specified in the
-      `command` field.
-    - Default size is 1x1 if not specified.
-    """
-
-    type: str
-    location: Location
-    size: Size | None = None
-    icon: str | None = None
-    text: str | None = None
-    command: EntityCommand | None = None
-
-
-def create_ui_text(
-    text: str,
-    x: int,
-    y: int,
-    size: Size | None = None,
-    cmd: str | EntityCommand | None = None,
-) -> UiItem:
-    """
-    Create a text UI item.
-
-    :param text: the text to show in the UI item.
-    :param x: x-position, 0-based.
-    :param y: y-position, 0-based.
-    :param size: item size, defaults to 1 x 1 if not specified.
-    :param cmd: associated command to the text item. A string parameter corresponds to
-                a simple command, whereas an ``EntityCommand`` allows to customize the
-                command for example with number of repeats.
-    :return: the created UiItem
-    """
-    if isinstance(cmd, str):
-        cmd = EntityCommand(cmd)
-    return UiItem("text", Location(x, y), size=size, text=text, command=cmd)
-
-
-def create_ui_icon(
-    icon: str,
-    x: int,
-    y: int,
-    size: Size | None = None,
-    cmd: str | EntityCommand | None = None,
-) -> UiItem:
-    """
-    Create an icon UI item.
-
-    The icon identifier consists of a prefix and a resource identifier,
-    separated by `:`. Available prefixes:
-      - `uc:` - integrated icon font
-      - `custom:` - custom resource
-
-    :param icon: the icon identifier of the icon to show in the UI item.
-    :param x: x-position, 0-based.
-    :param y: y-position, 0-based.
-    :param size: item size, defaults to 1 x 1 if not specified.
-    :param cmd: associated command to the text item. A string parameter corresponds to
-                a simple command, whereas an ``EntityCommand`` allows to customize the
-                command for example with number of repeats.
-    :return: the created UiItem
-    """
-    if isinstance(cmd, str):
-        cmd = EntityCommand(cmd)
-    return UiItem("icon", Location(x, y), size=size, icon=icon, command=cmd)
-
-
-@dataclass
-class UiPage:
-    """
-    Definition of a complete user interface page.
-
-    Default grid size is 4x6 if not specified.
-    """
-
-    page_id: str
-    name: str
-    _: KW_ONLY
-    grid: Size = None
-    items: list[UiItem] = None
-
-    def __post_init__(self):
-        """Post initialization to set required fields."""
-        # grid and items are required Integration-API fields
-        if self.grid is None:
-            self.grid = Size(4, 6)
-        if self.items is None:
-            self.items = []
-
-    def add(self, item: UiItem):
-        """Append the given UiItem to the page items."""
-        self.items.append(item)
+def _list_items_asdict(obj: list[Any]):
+    """Convert a list with (mixed) dataclass items to dictionary mapping items."""
+    return list(
+        map(
+            lambda item: (
+                dataclasses.asdict(item) if dataclasses.is_dataclass(item) else item
+            ),
+            obj,
+        )
+    )
 
 
 class Remote(Entity):
@@ -294,8 +128,8 @@ class Remote(Entity):
         features: list[Features],
         attributes: dict[str, Any],
         simple_commands: list[str] | None = None,
-        button_mapping: list[DeviceButtonMapping] | None = None,
-        ui_pages: list[UiPage] | None = None,
+        button_mapping: list[DeviceButtonMapping | dict[str, Any]] | None = None,
+        ui_pages: list[UiPage | dict[str, Any]] | None = None,
         area: str | None = None,
         cmd_handler: CommandHandler = None,
     ):
@@ -308,7 +142,9 @@ class Remote(Entity):
         :param attributes: remote attributes
         :param simple_commands: optional list of supported remote command identifiers
         :param button_mapping: optional command mapping of physical buttons
-        :param ui_pages: optional user interface page definitions
+               Either with DeviceButtonMapping items or plain dictionary items.
+        :param ui_pages: optional user interface page definitions.
+               Either with UiPage items or plain dictionary items.
         :param area: optional area
         :param cmd_handler: handler for entity commands
         """
@@ -316,9 +152,9 @@ class Remote(Entity):
         if simple_commands:
             options["simple_commands"] = simple_commands
         if button_mapping:
-            options["button_mapping"] = button_mapping
+            options["button_mapping"] = _list_items_asdict(button_mapping)
         if ui_pages:
-            options["user_interface"] = {"pages": ui_pages}
+            options["user_interface"] = {"pages": _list_items_asdict(ui_pages)}
         super().__init__(
             identifier,
             name,
